+++ src/components/VerticalImgCard.tsx
